+++ src/types.ts
@@ -696,6 +696,18 @@ export interface KaboomCtx {
 		curHeight: number,
 	) => void): void,
 	/**
+	 * Register an event that runs when a gamepad is connected.
+	 * 
+	 * @since v3000.0
+	 */
+	onGamepadConnect(action: (gamepad: Gamepad) => void): void,
+	/**
+	 * Register an event that runs when a gamepad is disconnected.
+	 * 
+	 * @since v3000.0
+	 */
+	onGamepadDisconnect(action: (gamepad: Gamepad) => void): void,
+	/**
 	 * Register an event that runs once when 2 game objs with certain tags collides (required to have area() component).
 	 *
 	 * @since v2000.1
@@ -935,12 +947,48 @@ export interface KaboomCtx {
 	 * @since v3000.0
 	 */
 	onVirtualButtonRelease(btn: VirtualButton, action: () => void): EventController,
-	onGamepadButtonDown(btn: string, action: (btn: string) => void): EventController,
-	onGamepadButtonDown(action: (btn) => void): EventController,
-	onGamepadButtonPress(btn: string, action: (btn: string) => void): EventController,
-	onGamepadButtonPress(action: (btn) => void): EventController,
-	onGamepadButtonRelease(btn: string, action: (btn: string) => void): EventController,
-	onGamepadButtonRelease(action: (btn) => void): EventController,
+	/**
+	 * Register an event that runs every frame when certain gamepad button is held down.
+	 * 
+	 * @since v3000.0
+	 */
+	onGamepadButtonDown(btn: GamepadButton, action: (btn: GamepadButton) => void): EventController,
+	/**
+	 * Register an event that runs every frame when any gamepad button is held down.
+	 * 
+	 * @since v3000.0
+	 */
+	onGamepadButtonDown(action: (btn: GamepadButton) => GamepadButton): EventController,
+	/**
+	 * Register an event that runs when user presses certain gamepad button.
+	 * 
+	 * @since v3000.0
+	 */
+	onGamepadButtonPress(btn: GamepadButton, action: (btn: GamepadButton) => void): EventController,
+	/**
+	 * Register an event that runs when user presses any gamepad button.
+	 * 
+	 * @since v3000.0
+	 */
+	onGamepadButtonPress(action: (btn: GamepadButton) => GamepadButton): EventController,
+	/**
+	 * Register an event that runs when user releases any gamepad button.
+	 *
+	 * @since v3000.0
+	 */
+	onGamepadButtonRelease(btn: GamepadButton, action: (btn: GamepadButton) => void): EventController,
+	/**
+	 * Register an event that runs when user releases certain gamepad button.
+	 *
+	 * @since v3000.0
+	 */
+	onGamepadButtonRelease(action: (btn: GamepadButton) => void): EventController,
+	/**
+	 * Register an event that runs when the gamepad axis exists.
+	 * 
+	 * @since v3000.0
+	 */
+	onGamepadStick(stick: "left" | "right", action: (value: Vec2) => void): EventController,
 	/**
 	 * Sets the root for all subsequent resource urls.
 	 *
@@ -1336,6 +1384,24 @@ export interface KaboomCtx {
 	 */
 	isVirtualButtonReleased(btn: VirtualButton): boolean,
 	/**
+	 * If a gamepad button is just pressed last frame
+	 *
+	 * @since v3000.0
+	 */
+	isGamepadButtonPressed(btn?: GamepadButton): boolean,
+	/**
+	 * If a gamepad button is currently held down.
+	 *
+	 * @since v3000.0
+	 */
+	isGamepadButtonDown(btn?: GamepadButton): boolean,
+	/**
+	 * If a gamepad button is just released last frame.
+	 *
+	 * @since v3000.0
+	 */
+	isGamepadButtonReleased(btn?: GamepadButton): boolean,
+	/**
 	 * List of characters inputted since last frame.
 	 *
 	 * @since v3000.0
@@ -1389,6 +1455,12 @@ export interface KaboomCtx {
 	 * Get / set gravity.
 	 */
 	gravity(g?: number): number,
+	/**
+	 * Get connected gamepads.
+	 * 
+	 * @since v3000.0
+	 */
+	getGamepads(): Gamepad[],
 	/**
 	 * Get / set the cursor (css). Cursor will be reset to "default" every frame so use this in an per-frame action.
 	 *
@@ -2208,6 +2280,24 @@ export type MouseButton =
 	| "middle"
 	| "back"
 	| "forward"
+
+export type GamepadButton = 
+	| "north"
+	| "east"
+	| "south"
+	| "west"
+	| "ltrigger"
+	| "rtrigger"
+    | "lshoulder"
+    | "rshoulder"
+	| "select"
+    | "start"
+	| "lstick"
+	| "rstick"
+	| "dpad-north"
+    | "dpad-east"
+	| "dpad-south"
+    | "dpad-west"
 
 /**
  * Inspect info for a character.