+++ CHANGELOG.md
@@ -253,6 +253,12 @@ music.loop = true
 - added `onVirtualButtonPress()`, `onVirtualButtonDown()`, `onVirtualButtonRelease()`
 - fixed touches not treated as mouse
 - (**BREAK**) changed `onTouchStart()`, `onTouchMove()` and `onTouchEnd()` callback signature to `(pos: Vec2, touch: Touch) => void` (exposes the native `Touch` object)
+- added `onGamepadButtonPress()`, `onGamepadButtonDown()`, `onGamepadButtonRelease()`
+- added `isGamepadButtonPressed()`, `isGamepadButtonDown()`, `isGamepadButtonReleased()`
+- added `onGamepadStick()` to handle gamepad axes info for left and right sticks
+- added `getConnectedGamepads()`
+- added `onGamepadConnect()` and `onGamepadDisconnect()`
+
 
 ## Level